+++ CHANGES.rst
@@ -11,6 +11,7 @@ Unreleased Changes
 * `Issue #86 <https://github.com/jantman/biweeklybudget/issues/86>`_ - Allow budget transfers to periodic budgets.
 * `Issue #74 <https://github.com/jantman/biweeklybudget/issues/74>`_ - Warning notification for low balance should take current pay period's overall allocated sum, minus reconciled transactions, into account.
 * Fix some template bugs that were causing HTML to be escaped into plaintext.
+* `Issue #15 <https://github.com/jantman/biweeklybudget/issues/15>`_ - Add pay period totals table to index page.
 
 0.1.2 (2017-05-28)
 ------------------